requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Pydantic
pydantic
pydantic-ai-slim[google]
pydantic-settings

# AI/ML
google-generativeai==0.8.3

# Database & Storage
supabase==2.9.0
postgrest==0.17.2

# Logging & Monitoring
logfire
logfire[fastapi]


# Image Processing
Pillow

# Environment & Utils
python-dotenv==1.0.1

# Optional: Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==4.1.0
pytest-mock==3.12.0

